pyproject.toml
[tool.poetry]
name = "weather-data-streamin"
version = "0.1.0"
description = ""
authors = ["Sviataslau Brezouski <sviataslau.brezouski@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
kafka-python-ng = "^2.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
